DEV-1459 Hybrid model using both API and kubectl
There is a deficiency in the Google API client such that it
does not seem possible (or at least isn't straightforward) to
get a cluster connection directly as the token is not set
on the returned cluster object when it is created.

The new approach is to use the API to create the cluster, poll
until it is complete, use the `gcloud` Kubernetes connector
and `kubectl` to initialise ~/.kube/config properly, then
create the Fabric8 client using the no-argument default
constructor which uses the contents of the Kubernetes config.

This code works but is a bit rough and needs some attention
before it is production-ready.