DEV-1473 Add ProcessRunner test
Also some more testing for the KubernetesClusterProvider.

Despite a few attempts I am not adding any test for the
KubernetesCluster at this point because they will either
be long chains of mocks, or depend upon the creation of
a local environment. I thought maybe Acquillian might help
with this but it doesn't work the way I thought I might
so leaving it for now.