DEV-1864 Overhaul of platinum configuration
Originally we let the pipeline5 config bleed out into platinum as a convenience, but its proven to be a little
inconvenient for users.

Here we introduce a fit for purpose model for platinum:
- Support yaml for better readability
- Remove all redundant or unnecessary fields
- Add the ability to group multiple tumors with the same normal
- Name pairs tumor/normal over tumor/reference

The platinum model gets serialized into the pipeline model before its added to the config map. The multi-tumor samples
are split into multiple pipelines. The run name is used to avoid name clashes.

There is also a little trick with the setId where we append `-set` to it. This will avoid the situation where the Sample
name is the same as the tumor and we delete the sample on cleanup.