DEV-2715 Fix JAR path for dev flow
Stef pointed out the provided path would no longer work for in-place use
on a development machine. I've inverted the change from last week so
that the path to the JAR will always be `-local-SNAPSHOT` and the
Dockerfile will copy it to there.

Also cleaned up some stale imports in a class I happened to have open
when I started.