DEV-3118 Constant job count scheduler
Splits the "scheduling" and "submitting" that was formerly done by the
JobScheduler class into an interface (the scheduling) and the new JobSubmitter
class.

The original scheduling algorithm has been pulled into the TimedBatchScheduler
class, while the new ConstantJobCountScheduler implements an alternative way
that maintains the configured number of jobs running at once.

This new way means we don't have to guess or experimentally determine an
optimal batch submission delay, nor do we have to try to balance the count
of jobs against stability of the cluster. I used it for the pipeline 5.31
rerun and have worked out all the bugs in the implementation, and it really
helped by allowing me to maximise throughput without having to micro-tune the
batch submission delay. Also this implementation doesn't create any jobs until
it is ready to start them, which seems to help with the overall stability of
the cluster. I have noticed that autoscaling in particular works very well
using this setup, while in the past it hasn't really performed.

There is still some work to do here, a couple of the tests have to be fixed
and the configuration parsing needs to be completed (it currently is hardcoded
to only use the new implementation, and only with a hardcoded active job
count). Want to get some eyes on it first and if no major issues are raised
then I'll make those changes.