DEV-3228 Improve handling of multi-set samples
Prior to this change if we ever discovered multiple sets for the same sample
we would be unable to continue. However this situation does occur when a new
reference sample is provided and the tumor sample re-used.

Use the status of the attached runs when this situation is encountered to try
to resolve the correct set.