DEV-3228 Improve handling of multi-set samples (#38)
* DEV-3228 Improve handling of multi-set samples

Prior to this change if we ever discovered multiple sets for the same sample
we would be unable to continue. However this situation does occur when a new
reference sample is provided and the tumor sample re-used.

Use the status of the attached runs when this situation is encountered to try
to resolve the correct set.

* DEV-3228 Speed up start of reruns

Large rerun sets take a long time to get going because the API has to
resolve or create all the runs first. Do this with a parallelStream as a
quick and cheap way to improve the experience.

* DEV-3228 Leverage API canonical set resolution

This was just added to the API and greatly simplifies the code here.

* DEV-3228 Fix up ApiRerunTest

---------

Co-authored-by: Ned Leitch <ned.leitch@gmail.com>