DEV-3118 Refactor to improve testability
The scheduling code has been moved to its own package, the Delay class has been
restructured to allow it to be used in more circumstances, the unit tests have
been repaired and both scheduling models are supported in the configuration.

Next steps could be to leverage the changes to the Delay class to improve the
unit testing.