DEV-3324 Make Kubernetes client calls from proxy
Added the remaining calls to the proxy as not all of them had been moved over
yet.

Updated the `platinum` script to remove the thread pooling override as this
was a temporary measure. I think I will actually just leave that out as there
are more changes coming to the order in which things are done.

Remove unneeded parameters from the JobSubmitter class.

Add ConstantJobCountSchedulerTest, refactoring slightly to facilitate.