DEV-3324 Add client proxy to allow reconnects (#42)
* DEV-3324 Add client proxy to allow reconnects

Add the KubernetesClientProxy class so that the several new places from which
there may now be Kubernetes client exceptions thrown can request a client
re-authorise rather than failing that sample's run or throwing the exception
all the way back up.

Also start breaking up the massive `kubernetes` pakcage which did not really
have just Kubernetes-related stuff in it.

Lastly override the default threading as the application was taking longer than
acceptable to run in Kubernetes. This will have to be improved in a more
robust way; this is a temporary improvement to get the 5.32 rerun done.

* DEV-3324 Make Kubernetes client calls from proxy

Added the remaining calls to the proxy as not all of them had been moved over
yet.

Updated the `platinum` script to remove the thread pooling override as this
was a temporary measure. I think I will actually just leave that out as there
are more changes coming to the order in which things are done.

Remove unneeded parameters from the JobSubmitter class.

Add ConstantJobCountSchedulerTest, refactoring slightly to facilitate.

---------

Co-authored-by: Ned Leitch <ned.leitch@gmail.com>