DEV-3323 Process (on both ends) samples in order
The advantage to be derived from the use of futures is essentially limited to
reruns as in this case we have several thousand samples involved in the typical
case.

The intention of the design that has been adopted is that an executor service
will populate the map of `PipelineInputs`, keyed on sample name, asynchronously
to the actual creation of the pipeline jobs that will use the inputs. That
creation of the pipeline jobs iterates over the inputs, calling `get()` on the
future that will contain the `PipelineInput`.

Without some sort of ordering, the asynchronous population thread could, in the
worst case, populate all the other `PipelineInput` instances before the very
first one that the pipeline job creation thread tried to access, leading to no
improvement from where we are today.

Thus this change introduces a protocol on both ends. The asynchronous thread
will sort the sample set as it is populating the map, as will the consuming
thread as it processes the list. In this way the samples that sort to the start
of the list should have futures for which a `.get()` will return quickly, and
thus be ready for processing by the second thread sooner, thus providing an
advantage over the status quo.