DEV-3323 Populate PDL asynchronously (#45)
* DEV-3323 Pipeline-specific k8 code to subpackage

I was finding the number of classes in the original package to be just too
many and this seemed a natural split.

* DEV-3323 Remove duplicated authorisation block

There was a block of duplicate code in the KubernetesEngine that was removed
and I also renamed the proxy's `reAuthorise` method to just plain `authorise`
since it sometimes will be the first invocation.

* DEV-3323 Invert jobIs helper method logic

Intellij keeps complaining about this method as always being inverted when it
is called so I just fixed it.

* DEV-3323 Defer PipelineInput creation

This is the majority of the implementation for this story, just build up
futures to contain the PipelineInput objects rather than storing them
directly. This also necessitated a move to changing the PDLConversion
away from outputting a List of inputs to a Map keyed by the sample names,
since we use the sample for some of the other processing between the PDL
creation and the actual submission of the jobs, and can no longer just
pull it out of the PDL.

There is still some work to do here. The testing of the PDL conversions is
lacking and should be improved. The executor services should maybe be injected
into the PDLConversion implementations, though they could also be considered
an implementation detail.

Also there needs to be some consideration to the order in which items are
processed in the map of pipeline inputs, since they will be processed asynch-
ronously. If the code that needs the completed PipelineInput values does not
wait on them in the same order they are being populated in then there is a
very real risk of starvation as the `.get()` on a Future waits until (in the
worst case) all of the inputs are processed. So we need to sort on the way in
and out in some consistent way. I do not feel the rest of the testing is up
to measuring yet whether this is happening.

* DEV-3323 Process (on both ends) samples in order

The advantage to be derived from the use of futures is essentially limited to
reruns as in this case we have several thousand samples involved in the typical
case.

The intention of the design that has been adopted is that an executor service
will populate the map of `PipelineInputs`, keyed on sample name, asynchronously
to the actual creation of the pipeline jobs that will use the inputs. That
creation of the pipeline jobs iterates over the inputs, calling `get()` on the
future that will contain the `PipelineInput`.

Without some sort of ordering, the asynchronous population thread could, in the
worst case, populate all the other `PipelineInput` instances before the very
first one that the pipeline job creation thread tried to access, leading to no
improvement from where we are today.

Thus this change introduces a protocol on both ends. The asynchronous thread
will sort the sample set as it is populating the map, as will the consuming
thread as it processes the list. In this way the samples that sort to the start
of the list should have futures for which a `.get()` will return quickly, and
thus be ready for processing by the second thread sooner, thus providing an
advantage over the status quo.

* DEV-3323 Using a supplier instead. (#46)

* DEV-3323 Fix up tests for Supplier version

---------

Co-authored-by: Ned Leitch <ned.leitch@gmail.com>
Co-authored-by: pauldwolfe <p.wolfe@hartwigmedicalfoundation.nl>