DEV-3337 Fix auth loop
Platinum would sit forever claiming it was "Authorising with cluster" but never
getting past that.

The cause was incorrect configuration being pushed which was causing a
KubernetesClientException, which itself was tricking the code into thinking
it needed to re-authorise.

I modified the code to sanitise the names of a few entries that had recently
changed and which weren't being converted to the DNS-compliant names Kubernetes
requires.

Test cases should also be written to address this. For the moment just pushing
implementation updates so others aren't blocked.