DEV-3378 Fix handling of existing jobs
We've got another situation like the recent "auth loop" that was previously
reported. This one is happening when a job already exists, as when the batch
was already executed once and now the user is trying again.

Similar fix, this was just a bug in when the job name was being converted to
Kubernetes-compatible format.