DEV-3396 Improve job status detection
The previous incarnation of this was confusing and not necessarily without bugs
but this version using the Conditions attribute of the job status is much more
clean than trying to infer the completion status from the state of the pods as
before.

Tests still need to be fixed, the next commit will have those completed.