DEV-3487 Default to huge batch
Aside from in the reruns it is probably the case the batch size is something
the user just doesn't want to fuss about. Just default to the maximum integer
for the batch size if it is not provided in the configuration.

Remove the default size from the "active jobs" list in the
`ConstantJobCountScheduler` implementation as this was just an optimisation
and will cause problems if the size is very large.