DEV-3670 Add in missing authorise for cluster
Without this Platinum just won't work unless there is an active and
correct cluster configuration already.

I think in the past we've just authorised manually maybe. I know this
functionality worked at some point but it is possible it hasn't worked
correctly since the introduction of the Kubernetes client proxy.