DEV-3670 Disallow confusing API override
Prior to this change there were two places a URL for the API could be
specified; one in the `hmfConfiguration` element and the other in the
argument overrides.

This is quite confusing. The presence of the optional API URL in the
`hmfConfiguration` is used by Platinum to make decisions about how to
proceed, while the presence of the URL in the overrides is passed to
the pipeline and used by it to make decisions about how to behave.
The upshot is that essentially they are not compatible with each other
and the only way I can see to provide proper semantics from Platinum in
line with supporting the required use cases is to disallow the use of
the override from Platinum.

I chose a fast failure rather than implicitly ignoring the override for
somewhat obvious reasons.