DEV-3670 Deprecate time-based scheduling
The time-based scheduler uses a batch size and delay, and submits a
batch of `size` pipelines every `delay` minutes. It was the original
batch scheduler in Platinum but suffers from some shortcomings:

* The user needs to be able to estimate the size of the batch and delay
  that should be configured before running anything;
* No throttling of requests to Kubernetes is done which can result in
  overwhelming the cluster and instability;
* It doesn't have any intelligence around whether it should continue
  submitting jobs according to the schedule.

The constant-size scheduler has been in use for some time and solves all
of these problems. It submits only the specified number of jobs at a
time rather than pushing all the jobs to Kubernetes at once and trying
to manage their run state. This allows it to be equally "gentle" on the
Kubernetes cluster whether it is running batches of 10 or 10000.

Update the README accordingly and deprecate the time-based scheduler for
later removal.

Also fix up the line breaks on the README as it was getting hard to edit
and they should be limited to some reasonable size, I selected the
somewhat-accepted 132 characters.