DEV-3670 Make upgrade branch-aware
This is the first step in managing versions properly so we never again get
ourselves into a position where users cannot run the latest Pipeline5 and
Platinum code together. Recently users have been getting stuck without a
working Platinum since it is not version aware and has also not been updated
to work with Pipeline5 5.33.

These changes will not fix the problem by themselves but they give us a
mechanism for providing more thoughtful releases. I have begun code changes
that allow Platinum to detect whether it is being asked to run against an
incompatible pipeline version which should help too. The implementation in
the script is quite a simple wrapper around `git` but with the added basic
functionality of writing the current version to a dotfile before doing the
update to help people get back to where they were if needed. I have not
yet decided whether to provide an automated "rollback" capability using
this dotfile's contents.

I have also updated the README.md to set expectations around the capabilities
of the `platinum` script and clue the user in to the fact it is just using
`git` and other tools under the covers. This should help avoid some of the
issues we've had recently where internal users weren't really clear on what
they could do to help themselves; once the proper branch structure is in place
I expect this is to be a good starting point.

Also cut out some cruft from the README.md where easy, I think it still has
some cleanup to be done but trying to avoid a whole-hog approach.