DEV-3670 Check image against supported versions
Proactively check the requested image in the YAML against the versions of the
pipeline supported by this Platinum and bail out if the user has requested an
untenable situation.

I'm trying to get the balance right here between overengineering a bulletproof
solution and what we have now, which is no real agreement between the Platinum
and Pipeline5 sides.

Believe this is a good place to start but there are limitations to this method
that are worth identifying already:

* Agreement between the requested Docker image and the GCP disk image is not
  checked. I think doing so would be overstepping because users may have
  legitimate reasons for doing that and I would not want to get in the way of
  experimental work.
* I've specifically chosen to hardcode the checks so that the compatible
  range is tied directly to the code itself. Some people might prefer a config
  or other abstraction but I think this encourages deliberate and thoughtful
  change by somebody who knows for sure about the compatibility of a pipeline
  and a Platinum version (the developer involved with one or the other).