DEV-3670 Sundry improvements to Platinum (#56)
* DEV-3670 Add in missing authorise for cluster

Without this Platinum just won't work unless there is an active and
correct cluster configuration already.

I think in the past we've just authorised manually maybe. I know this
functionality worked at some point but it is possible it hasn't worked
correctly since the introduction of the Kubernetes client proxy.

* DEV-3670 Disallow confusing API override

Prior to this change there were two places a URL for the API could be
specified; one in the `hmfConfiguration` element and the other in the
argument overrides.

This is quite confusing. The presence of the optional API URL in the
`hmfConfiguration` is used by Platinum to make decisions about how to
proceed, while the presence of the URL in the overrides is passed to
the pipeline and used by it to make decisions about how to behave.
The upshot is that essentially they are not compatible with each other
and the only way I can see to provide proper semantics from Platinum in
line with supporting the required use cases is to disallow the use of
the override from Platinum.

I chose a fast failure rather than implicitly ignoring the override for
somewhat obvious reasons.

* DEV-3670 Deprecate time-based scheduling

The time-based scheduler uses a batch size and delay, and submits a
batch of `size` pipelines every `delay` minutes. It was the original
batch scheduler in Platinum but suffers from some shortcomings:

* The user needs to be able to estimate the size of the batch and delay
  that should be configured before running anything;
* No throttling of requests to Kubernetes is done which can result in
  overwhelming the cluster and instability;
* It doesn't have any intelligence around whether it should continue
  submitting jobs according to the schedule.

The constant-size scheduler has been in use for some time and solves all
of these problems. It submits only the specified number of jobs at a
time rather than pushing all the jobs to Kubernetes at once and trying
to manage their run state. This allows it to be equally "gentle" on the
Kubernetes cluster whether it is running batches of 10 or 10000.

Update the README accordingly and deprecate the time-based scheduler for
later removal.

Also fix up the line breaks on the README as it was getting hard to edit
and they should be limited to some reasonable size, I selected the
somewhat-accepted 132 characters.

* DEV-3670 Make upgrade branch-aware

This is the first step in managing versions properly so we never again get
ourselves into a position where users cannot run the latest Pipeline5 and
Platinum code together. Recently users have been getting stuck without a
working Platinum since it is not version aware and has also not been updated
to work with Pipeline5 5.33.

These changes will not fix the problem by themselves but they give us a
mechanism for providing more thoughtful releases. I have begun code changes
that allow Platinum to detect whether it is being asked to run against an
incompatible pipeline version which should help too. The implementation in
the script is quite a simple wrapper around `git` but with the added basic
functionality of writing the current version to a dotfile before doing the
update to help people get back to where they were if needed. I have not
yet decided whether to provide an automated "rollback" capability using
this dotfile's contents.

I have also updated the README.md to set expectations around the capabilities
of the `platinum` script and clue the user in to the fact it is just using
`git` and other tools under the covers. This should help avoid some of the
issues we've had recently where internal users weren't really clear on what
they could do to help themselves; once the proper branch structure is in place
I expect this is to be a good starting point.

Also cut out some cruft from the README.md where easy, I think it still has
some cleanup to be done but trying to avoid a whole-hog approach.

* DEV-3670 Check image against supported versions

Proactively check the requested image in the YAML against the versions of the
pipeline supported by this Platinum and bail out if the user has requested an
untenable situation.

I'm trying to get the balance right here between overengineering a bulletproof
solution and what we have now, which is no real agreement between the Platinum
and Pipeline5 sides.

Believe this is a good place to start but there are limitations to this method
that are worth identifying already:

* Agreement between the requested Docker image and the GCP disk image is not
  checked. I think doing so would be overstepping because users may have
  legitimate reasons for doing that and I would not want to get in the way of
  experimental work.
* I've specifically chosen to hardcode the checks so that the compatible
  range is tied directly to the code itself. Some people might prefer a config
  or other abstraction but I think this encourages deliberate and thoughtful
  change by somebody who knows for sure about the compatibility of a pipeline
  and a Platinum version (the developer involved with one or the other).

* DEV-3670 Simplify test

Just make all calls to the pipeline version extractor return the same
string they were called with.

* DEV-3670 Wire in current version check

* DEV-3670 Add Kubernetes refresher to README

* DEV-3670 Address PR comments

---------

Co-authored-by: Ned Leitch <ned.leitch@gmail.com>