INFRA-344 Pass rerun API URL to Pv5
While it is not allowed to directly override the URL for the
HMF API in the Platinum configuration, for the reruns we do
want a URL to be set.

Just pass it through directly if a rerun is happening. I HAVE NOT
tested this as we have a rerun in progress already but expect it
might be all that is required.